refactor: Model 구조체가 CurrentWeather, FiveDaysWeather를 갖는 구조로 변경
